+++ src/FluxFlow.Components.Resources/Contracts/ResourceDiagnosticSeverity.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.Resources.Contracts;
+
+public enum ResourceDiagnosticSeverity
+{
+    Information,
+    Warning,
+    Error
+}